move impl to StatFunctions and add revise doc
